requirements.txt
# AI Crypto Investment Intelligence Platform - Dependencies
# Core
fastapi>=0.109.2
uvicorn>=0.27.1
pydantic>=2.6.1
email-validator>=2.3.0

# Data Processing
pandas>=2.2.0
numpy>=1.26.4

# Machine Learning
scikit-learn>=1.4.0
torch>=2.2.0
xgboost>=2.0.0
lightgbm>=4.3.0
yfinance>=0.2.36

# Database
motor>=3.3.2
pymongo>=4.6.1

# Real-Time
python-socketio>=5.11.0

# Visualization & Dashboard
plotly>=5.18.0
matplotlib>=3.8.3

# HTTP Requests
requests>=2.31.0

# Scientific Computing
scipy>=1.12.0

# Authentication
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.1.2

# Reports
fpdf2>=2.7.7
python-dotenv>=1.0.1

# Phase 1-4 Advanced Features
vaderSentiment>=3.3.2
textblob>=0.17.1
firebase-admin>=6.4.0
httpx>=0.27.0
slowapi>=0.1.9
aioredis>=2.0.1
redis>=5.0.1
pydantic-settings>=2.1.0
nltk>=3.8.1
